temp = fake news

+++ src/lcd/lcd.h
@@ -34,7 +34,7 @@ int lcdGetDescaleCycle(void);
 
 void lcdSetDescaleCycle(int cycle);
 void lcdSetPressure(float val);
-void lcdSetTemperature(int val);
+void lcdSetTemperature(uint16_t val);
 void lcdTargetState(int val);
 void lcdSetWeight(float val);
 void lcdSetFlow(int val);
